+++ src/Graph/Nodes/Hypercube.php
@@ -12,4 +12,11 @@ interface Hypercube extends Node
      * @return \Traversable<array>
      */
     public function sides() : Traversable;
+
+    /**
+     * Does the hypercube reduce to a single point?
+     *
+     * @return bool
+     */
+    public function isPoint() : bool;
 }